review documentation

+++ src/controller/psicologosController.js
@@ -35,7 +35,7 @@ const psicologosController = {
             const { id } = req.params;
             const psicologoID = await Psicologos.findByPk(id);
             if (!psicologoID) {
-                res.status(404).json("Não há um psicólogo com este id " + id);
+                res.status(404).json("Id não encontrado");
             }
             res.status(200).json(psicologoID);
         }
@@ -47,16 +47,24 @@ const psicologosController = {
     async deletePsychologist (req, res) {
         try {
             const { id } = req.params;
-            const psicologoID = await Psicologos.findByPk(id);
-            await Psicologos.destroy({
-                where: {
-                id,
-                },
+            const psicologoID = await Psicologos.findByPk(id, {
+                include: {
+                    model: Atendimentos,
+                }
             });
             if (!psicologoID) {
-                return res.status(404).json("Não há um psicólogo com esse ID" + id);}
-
-            res.sendStatus(204);
+                return res.status(404).json("Id não encontrado");}
+            if(psicologoID.atendimentos.length === 0) {
+                await Psicologos.destroy({
+                    where: {
+                    id,
+                    },
+                });
+                res.sendStatus(204);
+            }
+            else {
+                res.status(401).json('Cadastro de psicólogo não pode ser deletado pois possui atendimentos agendados.');
+            } 
         }
         catch (error) {
                 console.error(error);
@@ -70,16 +78,16 @@ const psicologosController = {
             const novaSenha = bcrypt.hashSync(senha, 10);
             const psicologoID = await Psicologos.findByPk(id);
             if (!psicologoID) {
-                res.status(404).json("Não há um psicólogo com este id " + id);
+                res.status(404).json("Id não encontrado");
             }
             else {
-            const atualizarPsicologo = await Psicologos.update({ nome, email, senha: novaSenha, apresentacao, },
+            await Psicologos.update({ nome, email, senha: novaSenha, apresentacao, },
             {
                 where: {
                 id,
                 },
             });
-            res.status(200).json(atualizarPsicologo)
+            res.status(200).json(psicologoID)
             }
         }
         catch (error) {
